🚚 Move Result objects in seperate namespace
